+++ lib/src/features/menu/bloc/blocs/top_level_blocs/category_bloc.dart
@@ -1,0 +1,25 @@
+import 'package:bloc/bloc.dart';
+import 'package:coffeeshop/src/features/menu/bloc/states/category_state.dart';
+import 'package:coffeeshop/src/features/menu/data/repositories/category_repository.dart';
+
+abstract class CategoryEvent {}
+
+class LoadCategoriesEvent extends CategoryEvent {
+  LoadCategoriesEvent();
+}
+
+class CategoryBloc extends Bloc<CategoryEvent, CategoryState> {
+  final CategoryRepository repository;
+
+  CategoryBloc(this.repository) : super(CategoryInit()) {
+    on<LoadCategoriesEvent>((event, emit) async {
+      emit(CategoryLoading());
+      try {
+        final categories = await repository.fetchCategories();
+        emit(CategoryLoaded(categories));
+      } catch (e) {
+        emit(CategoryError(e.toString()));
+      }
+    });
+  }
+}
