+++ lib/src/features/menu/data/DTO/product_dto.dart
@@ -1,0 +1,25 @@
+class ProductDTO {
+  final int id;
+  final String name;
+  final String image;
+  final double price;
+  final int categoryId;
+
+  ProductDTO({
+    required this.id,
+    required this.name,
+    required this.image,
+    required this.price,
+    required this.categoryId
+  });
+
+  factory ProductDTO.fromJson(Map<String, dynamic> json) {
+    return ProductDTO(
+      id: json['id'],
+      name: json['name'],
+      image: json['imageUrl'],
+      price: double.parse(json['prices'][0]['value']),
+      categoryId: json['category']['id']
+    );
+  }
+}
